+++ Tests/ReactiveSwiftTests/SignalSpec.swift
@@ -3093,6 +3093,18 @@ class SignalSpec: QuickSpec {
 				observer2.sendCompleted()
 			}
 		}
+
+		describe("promoteError") {
+			it("should infer the error type from the context") {
+				let combined: Any = Signal
+					.combineLatest(Signal<Int, NoError>.never.promoteError(),
+					               Signal<Double, TestError>.never,
+					               Signal<Float, NoError>.never.promoteError(),
+					               Signal<UInt, POSIXError>.never.flatMapError { _ in .empty })
+
+				expect(combined is Signal<(Int, Double, Float, UInt), TestError>) == true
+			}
+		}
 	}
 }